split(store/sqlite): break store_authz.go into three topic files
store_authz.go (657 lines) exceeded the 500-line target. Split into:
- store_authz_membership.go: AddUserToGroup, RemoveUserFromGroup,
  GrantUserPermission, RevokeUserPermission, GrantGroupPermission,
  RevokeGroupPermission (the FK-adjacent write operations)
- store_authz_masks.go: userEffectivePermissionPredicateSQL, effectiveMaskSQL,
  PermissionMasksForUserResource, EffectiveMask (mask computation)
- store_authz_listing.go: all four listing functions (UserAuthzResourcesList,
  GroupAuthzResourcesList, ResourceAuthzGroupsList, ResourceAuthzUsersList)
  and their SQL constants and helpers

No logic changes.
